+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core/PParams.hs
@@ -3,13 +3,13 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE EmptyCase #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE RankNTypes #-}
@@ -39,6 +39,7 @@ module Cardano.Ledger.Core.PParams (
   ppMaxBHSizeL,
   ppKeyDepositL,
   ppPoolDepositL,
+  ppPoolDepositCompactL,
   ppEMaxL,
   ppNOptL,
   ppA0L,
@@ -57,6 +58,7 @@ module Cardano.Ledger.Core.PParams (
   ppuMaxBHSizeL,
   ppuKeyDepositL,
   ppuPoolDepositL,
+  ppuPoolDepositCompactL,
   ppuEMaxL,
   ppuNOptL,
   ppuA0L,
@@ -89,9 +91,19 @@ import Cardano.Ledger.BaseTypes (
   UnitInterval,
   maybeToStrictMaybe,
  )
-import Cardano.Ledger.Binary
-import Cardano.Ledger.Binary.Coders
-import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Ledger.Binary (
+  DecCBOR (..),
+  EncCBOR (..),
+  FromCBOR (..),
+  ToCBOR (..),
+  decodeRecordNamed,
+  encodeListLen,
+  encodeMapLen,
+  encodeWord,
+ )
+import Cardano.Ledger.Binary.Coders (Decode (..), Field, decode, field, invalidField)
+import Cardano.Ledger.Coin (Coin (..), partialCompactCoinL)
+import Cardano.Ledger.Compactible (Compactible (..), partialCompactFL)
 import Cardano.Ledger.Core.Era (Era (..), PreviousEra, ProtVerAtMost, fromEraCBOR, toEraCBOR)
 import Cardano.Ledger.HKD (HKD, HKDApplicative, HKDFunctor (..), NoUpdate (..))
 import Cardano.Ledger.Plutus.ToPlutusData (ToPlutusData (..))
@@ -356,7 +368,7 @@ class
   hkdKeyDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)
 
   -- | The amount of a pool registration deposit
-  hkdPoolDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)
+  hkdPoolDepositCompactL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f (CompactForm Coin))
 
   -- | epoch bound on pool retirement
   hkdEMaxL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochInterval)
@@ -444,7 +456,11 @@ ppKeyDepositL = ppLensHKD . hkdKeyDepositL @era @Identity
 
 -- | The amount of a pool registration deposit
 ppPoolDepositL :: forall era. EraPParams era => Lens' (PParams era) Coin
-ppPoolDepositL = ppLensHKD . hkdPoolDepositL @era @Identity
+ppPoolDepositL = ppPoolDepositCompactL . partialCompactCoinL
+
+-- | The amount of a pool registration deposit in compacted form
+ppPoolDepositCompactL :: forall era. EraPParams era => Lens' (PParams era) (CompactForm Coin)
+ppPoolDepositCompactL = ppLensHKD . hkdPoolDepositCompactL @era @Identity
 
 -- | epoch bound on pool retirement
 ppEMaxL :: forall era. EraPParams era => Lens' (PParams era) EpochInterval
@@ -508,9 +524,16 @@ ppuMaxBHSizeL = ppuLensHKD . hkdMaxBHSizeL @era @StrictMaybe
 ppuKeyDepositL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)
 ppuKeyDepositL = ppuLensHKD . hkdKeyDepositL @era @StrictMaybe
 
--- | The amount of a pool registration deposit
-ppuPoolDepositL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)
-ppuPoolDepositL = ppuLensHKD . hkdPoolDepositL @era @StrictMaybe
+-- | The amount of a pool registration deposit. The value must be small enough
+-- to fit into a Word64.
+ppuPoolDepositL ::
+  forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)
+ppuPoolDepositL = ppuLensHKD . hkdPoolDepositCompactL @era @StrictMaybe . partialCompactFL
+
+-- | The amount of a pool registration deposit in compacted form
+ppuPoolDepositCompactL ::
+  forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe (CompactForm Coin))
+ppuPoolDepositCompactL = ppuLensHKD . hkdPoolDepositCompactL @era @StrictMaybe
 
 -- | epoch bound on pool retirement
 ppuEMaxL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)